EmbedRequestMixin & ClassifyRequestMixin
